fix introduction app
