+++ app/src/main/java/com/marlon/portalusuario/presentation/splash/ActivitySplash.kt
@@ -1,25 +1,23 @@
-package com.marlon.portalusuario.intro
+package com.marlon.portalusuario.presentation.splash
 
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
-import android.os.Build
 import android.os.Bundle
-import android.os.PowerManager
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.preference.PreferenceManager
 import com.marlon.portalusuario.data.preferences.AppPreferences
+import com.marlon.portalusuario.intro.IntroActivity
 import com.marlon.portalusuario.presentation.splash.screen.SplashScreen
 import com.marlon.portalusuario.ui.theme.PortalUsuarioTheme
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
-
 @AndroidEntryPoint
-class SplashActivity : ComponentActivity() {
+class ActivitySplash : ComponentActivity() {
     @Inject lateinit var preferences: AppPreferences
     private val sharedPreferences: SharedPreferences by lazy {
         getSharedPreferences("myPrefs", Context.MODE_PRIVATE)
@@ -37,7 +35,6 @@ class SplashActivity : ComponentActivity() {
             else -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM)
         }
 
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) manageBatteryConsumption()
         if (!sharedPreferences.getBoolean("isIntroOpened", false)) {
             startActivity(Intent(this, IntroActivity::class.java))
             finish()
@@ -47,18 +44,6 @@ class SplashActivity : ComponentActivity() {
             PortalUsuarioTheme(darkTheme = settings.getString("keynoche", "") == "oscuro") {
                 SplashScreen()
             }
-        }
-    }
-
-    private fun manageBatteryConsumption() {
-        val powerManager = getSystemService(Context.POWER_SERVICE) as PowerManager
-        if (!powerManager.isIgnoringBatteryOptimizations(packageName)) {
-            startActivity(
-                Intent().apply {
-                    action = android.provider.Settings.ACTION_IGNORE_BATTERY_OPTIMIZATION_SETTINGS
-                    data = android.net.Uri.parse("package:$packageName")
-                }
-            )
         }
     }
 }
