+++ app/src/main/java/com/marlon/portalusuario/presentation/splash/screen/SplashScreen.kt
@@ -47,8 +47,7 @@ fun SplashScreen(viewModel: SplashViewModel = hiltViewModel()) {
 
     Surface(
         modifier = Modifier
-            .fillMaxSize()
-            .background(Color(0xFFF5F5F5)),
+            .fillMaxSize(),
         color = MaterialTheme.colorScheme.background
     ) {
         Column(
